requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
mercadopago==2.2.3